abstract channel logic into its own service
